Add 64bit rate/ceil support for htb class
Htb class has already supported 64bit rate and ceil settings for times.
Now, in this patch, we grant this ability to libnl library.

+++ lib/route/tc.c
@@ -646,7 +646,7 @@ int rtnl_tc_str2stat(const char *name)
  * 
  * @return Required transmit time in micro seconds.
  */
-int rtnl_tc_calc_txtime(int bufsize, int rate)
+int rtnl_tc_calc_txtime(int bufsize, uint64_t rate)
 {
 	double tx_time_secs;
 	
@@ -669,7 +669,7 @@ int rtnl_tc_calc_txtime(int bufsize, int rate)
  *
  * @return Size of buffer in bytes.
  */
-int rtnl_tc_calc_bufsize(int txtime, int rate)
+int rtnl_tc_calc_bufsize(int txtime, uint64_t rate)
 {
 	double bufsize;
